requirements.txt
asgiref==3.3.4
backtrader==1.9.76.123
backtrader-plotting==1.1.0
beautifulsoup4==4.9.3
bokeh==2.0.2
certifi==2020.6.20
chardet==4.0.0
click==7.1.2
cycler==0.10.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.4
django-import-export==2.5.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==2.10
Jinja2==3.0.0
joblib==1.0.1
kiwisolver==1.2.0
lxml==4.6.3
markdown2==2.4.0
MarkupPy==1.14
MarkupSafe==2.0.0
matplotlib==3.2.2
mibian==0.1.3
nsepy==0.8
numpy==1.19.2
odfpy==1.4.1
openpyxl==3.0.7
packaging==20.9
pandas==1.1.2
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.4.1
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.2
seaborn==0.11.1
six==1.15.0
sklearn==0.0
soupsieve==2.2.1
sqlparse==0.4.1
sv==0.1
TA-Lib @ file:///C:/Users/Jay%20Patil/PycharmProjects/pythonProject/TA_Lib-0.4.19-cp37-cp37m-win_amd64.whl
tablib==3.0.0
threadpoolctl==2.1.0
tornado==6.1
typing-extensions==3.10.0.0
urllib3==1.26.4
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
